backend/requirements.txt
fastapi
uvicorn
youtube-transcript-api
google-genai
python-dotenv
sse-starlette
